Refactor project structure for better scalability
