+++ app/__init__.py
@@ -1,0 +1,20 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
+
+db = SQLAlchemy()
+migrate = Migrate()
+
+def create_app():
+    app = Flask(__name__)
+    app.config.from_object('app.config.Config')
+
+    db.init_app(app)
+    migrate.init_app(app, db)
+
+    with app.app_context():
+        from app import routes, models
+        routes.register_routes(app)  # Register routes
+        db.create_all()
+
+    return app
